change header file to long long

+++ stoplights/solve.h
@@ -7,6 +7,6 @@
 #include <string>
 #include <vector>
 
-std::string solve(std::vector<std::vector<graph_edge> > graph, int startIndex, int endIndex);
+std::string solve(std::vector<std::vector<graph_edge> > graph, long long startIndex, long long endIndex);
 
 #endif
